+++ src/codegen/compile_classical.ml
@@ -7366,19 +7366,19 @@ module MakeSerialization (Strm : Stream) = struct
         reserve env get_data_buf 8l ^^
         get_x ^^ Float.unbox env ^^
         G.i (Store {ty = F64Type; align = 0; offset = 0L; sz = None})
-      | Prim ((Int64|Nat64) as pty) ->
+      | Prim (Int64|Nat64 as pty) ->
         reserve env get_data_buf 8l ^^
         get_x ^^ BoxedWord64.unbox env pty ^^
         G.i (Store {ty = I64Type; align = 0; offset = 0L; sz = None})
-      | Prim ((Int32|Nat32) as ty) ->
+      | Prim (Int32|Nat32 as ty) ->
         write_word_32 env get_data_buf (get_x ^^ BoxedSmallWord.unbox env ty)
       | Prim Char ->
         write_word_32 env get_data_buf (get_x ^^ TaggedSmallWord.lsb_adjust_codepoint env)
-      | Prim ((Int16|Nat16) as ty) ->
+      | Prim (Int16|Nat16 as ty) ->
         reserve env get_data_buf 2l ^^
         get_x ^^ TaggedSmallWord.lsb_adjust ty ^^
         G.i (Store {ty = I32Type; align = 0; offset = 0L; sz = Some Wasm.Types.Pack16})
-      | Prim ((Int8|Nat8) as ty) ->
+      | Prim (Int8|Nat8 as ty) ->
         write_byte env get_data_buf (get_x ^^ TaggedSmallWord.lsb_adjust ty)
       | Prim Bool ->
         write_byte env get_data_buf get_x
@@ -7894,13 +7894,13 @@ module MakeSerialization (Strm : Stream) = struct
           ReadBuf.read_float64 env get_data_buf ^^
           Float.box env
         end
-      | Prim ((Int64|Nat64) as pty) ->
+      | Prim (Int64|Nat64 as pty) ->
         with_prim_typ t
         begin
           ReadBuf.read_word64 env get_data_buf ^^
           BoxedWord64.box env pty
         end
-      | Prim ((Int32|Nat32) as pty) ->
+      | Prim (Int32|Nat32 as pty) ->
         with_prim_typ t
         begin
           ReadBuf.read_word32 env get_data_buf ^^
@@ -7913,14 +7913,14 @@ module MakeSerialization (Strm : Stream) = struct
           TaggedSmallWord.check_and_msb_adjust_codepoint env ^^
           TaggedSmallWord.tag env Char
         end
-      | Prim ((Int16|Nat16) as ty) ->
+      | Prim (Int16|Nat16 as ty) ->
         with_prim_typ t
         begin
           ReadBuf.read_word16 env get_data_buf ^^
           TaggedSmallWord.msb_adjust ty ^^
           TaggedSmallWord.tag env ty
         end
-      | Prim ((Int8|Nat8) as ty) ->
+      | Prim (Int8|Nat8 as ty) ->
         with_prim_typ t
         begin
           ReadBuf.read_byte env get_data_buf ^^
@@ -9079,9 +9079,9 @@ module StackRep = struct
     match normalize t with
     | Prim Bool -> SR.bool
     | Prim (Nat | Int) -> Vanilla
-    | Prim ((Nat64 | Int64) as pty) -> UnboxedWord64 pty
-    | Prim ((Nat32 | Int32) as pty) -> UnboxedWord32 pty
-    | Prim ((Nat8 | Nat16 | Int8 | Int16 | Char) as pty) -> UnboxedWord32 pty
+    | Prim (Nat64 | Int64 as pty) -> UnboxedWord64 pty
+    | Prim (Nat32 | Int32 as pty) -> UnboxedWord32 pty
+    | Prim (Nat8 | Nat16 | Int8 | Int16 | Char as pty) -> UnboxedWord32 pty
     | Prim (Text | Blob | Principal) -> Vanilla
     | Prim Float -> UnboxedFloat64
     | Obj (Actor, _, _) -> Vanilla
@@ -9198,17 +9198,16 @@ module StackRep = struct
        BoxedWord64.unbox env pty
 
     (* TaggedSmallWord types *)
-    | UnboxedWord32 (Type.(Int8 | Nat8 | Int16 | Nat16 | Char) as pty), Vanilla ->
+    | UnboxedWord32 Type.(Int8 | Nat8 | Int16 | Nat16 | Char as pty), Vanilla ->
        TaggedSmallWord.tag env pty
-    | Vanilla, UnboxedWord32 (Type.(Nat8 | Int8 | Nat16 | Int16 | Char) as pty) ->
+    | Vanilla, UnboxedWord32 Type.(Nat8 | Int8 | Nat16 | Int16 | Char as pty) ->
        TaggedSmallWord.untag env pty
 
     (* BoxedSmallWord types *)
     | UnboxedWord32 pty, Vanilla ->
        assert Type.(pty = Nat32 || pty = Int32);
        BoxedSmallWord.box env pty
-    | Vanilla, UnboxedWord32 ((Type.Nat32 | Type.Int32) as pty) ->
-       assert Type.(pty = Nat32 || pty = Int32);
+    | Vanilla, UnboxedWord32 Type.(Nat32 | Int32 as pty) ->
        BoxedSmallWord.unbox env pty
 
     | UnboxedFloat64, Vanilla -> Float.box env
@@ -10194,9 +10193,9 @@ module AllocHow = struct
   let stackrep_of_type t =
     let open Type in
     match normalize t with
-    | Prim ((Nat32 | Int32 | Nat16 | Int16 | Nat8 | Int8 | Char) as pty) ->
+    | Prim (Nat32 | Int32 | Nat16 | Int16 | Nat8 | Int8 | Char as pty) ->
        SR.UnboxedWord32 pty
-    | Prim ((Nat64 | Int64) as pty) -> SR.UnboxedWord64 pty
+    | Prim (Nat64 | Int64 as pty) -> SR.UnboxedWord64 pty
     | Prim Float -> SR.UnboxedFloat64
     | _ -> SR.Vanilla
 
@@ -10425,7 +10424,7 @@ let compile_unop env t op =
         get_n ^^
         G.i (Binary (Wasm.Values.I64 I64Op.Sub))
       )
-  | NegOp, Type.(Prim ((Int8 | Int16 | Int32) as p)) ->
+  | NegOp, Type.(Prim (Int8 | Int16 | Int32 as p)) ->
     StackRep.of_type t, StackRep.of_type t,
     Func.share_code1 Func.Never env (prim_fun_name p "neg32_trap") ("n", I32Type) [I32Type] (fun env get_n ->
       get_n ^^
@@ -11967,7 +11966,7 @@ and compile_prim_invocation (env : E.t) ae p es at =
     Region.load_word8 env ^^
     TaggedSmallWord.msb_adjust ty
 
-  | OtherPrim (("regionStoreNat8" | "regionStoreInt8") as p), [e0; e1; e2] ->
+  | OtherPrim ("regionStoreNat8" | "regionStoreInt8" as p), [e0; e1; e2] ->
     let ty = Type.(if p = "regionStoreNat8" then Nat8 else Int8) in
     SR.unit,
     compile_exp_as env ae SR.Vanilla e0 ^^
@@ -11976,7 +11975,7 @@ and compile_prim_invocation (env : E.t) ae p es at =
     TaggedSmallWord.lsb_adjust ty ^^
     Region.store_word8 env
 
-  | OtherPrim (("regionLoadNat16" | "regionLoadInt16") as p), [e0; e1] ->
+  | OtherPrim ("regionLoadNat16" | "regionLoadInt16" as p), [e0; e1] ->
     let ty = Type.(if p = "regionLoadNat16" then Nat16 else Int16) in
     SR.UnboxedWord32 ty,
     compile_exp_as env ae SR.Vanilla e0 ^^
@@ -11984,7 +11983,7 @@ and compile_prim_invocation (env : E.t) ae p es at =
     Region.load_word16 env ^^
     TaggedSmallWord.msb_adjust ty
 
-  | OtherPrim (("regionStoreNat16" | "regionStoreInt16") as p), [e0; e1; e2] ->
+  | OtherPrim ("regionStoreNat16" | "regionStoreInt16" as p), [e0; e1; e2] ->
     SR.unit,
     let ty = Type.(if p = "regionStoreNat16" then Nat16 else Int16) in
     compile_exp_as env ae SR.Vanilla e0 ^^
@@ -11993,13 +11992,13 @@ and compile_prim_invocation (env : E.t) ae p es at =
     TaggedSmallWord.lsb_adjust ty ^^
     Region.store_word16 env
 
-  | OtherPrim (("regionLoadNat32" | "regionLoadInt32") as p), [e0; e1] ->
+  | OtherPrim ("regionLoadNat32" | "regionLoadInt32" as p), [e0; e1] ->
     SR.UnboxedWord32 Type.(if p = "regionLoadNat32" then Nat32 else Int32),
     compile_exp_as env ae SR.Vanilla e0 ^^
     compile_exp_as env ae (SR.UnboxedWord64 Type.Nat64) e1 ^^
     Region.load_word32 env
 
-  | OtherPrim (("regionStoreNat32" | "regionStoreInt32") as p), [e0; e1; e2] ->
+  | OtherPrim ("regionStoreNat32" | "regionStoreInt32" as p), [e0; e1; e2] ->
     SR.unit,
     compile_exp_as env ae SR.Vanilla e0 ^^
     compile_exp_as env ae (SR.UnboxedWord64 Type.Nat64) e1 ^^
@@ -12007,13 +12006,13 @@ and compile_prim_invocation (env : E.t) ae p es at =
       (SR.UnboxedWord32 Type.(if p = "regionStoreNat32" then Nat32 else Int32)) e2 ^^
     Region.store_word32 env
 
-  | OtherPrim (("regionLoadNat64" | "regionLoadInt64") as p), [e0; e1] ->
+  | OtherPrim ("regionLoadNat64" | "regionLoadInt64" as p), [e0; e1] ->
     (SR.UnboxedWord64 Type.(if p = "regionLoadNat64" then Nat64 else Int64)),
     compile_exp_as env ae SR.Vanilla e0 ^^
     compile_exp_as env ae (SR.UnboxedWord64 Type.Nat64) e1 ^^
     Region.load_word64 env
 
-  | OtherPrim (("regionStoreNat64" | "regionStoreInt64") as p), [e0; e1; e2] ->
+  | OtherPrim ("regionStoreNat64" | "regionStoreInt64" as p), [e0; e1; e2] ->
     SR.unit,
     compile_exp_as env ae SR.Vanilla e0 ^^
     compile_exp_as env ae (SR.UnboxedWord64 Type.Nat64) e1 ^^
@@ -12228,12 +12227,12 @@ and compile_prim_invocation (env : E.t) ae p es at =
   | OtherPrim ("arrayToBlob" | "arrayMutToBlob"), e ->
     const_sr SR.Vanilla (Arr.toBlob env)
 
-  | OtherPrim (("stableMemoryLoadNat32" | "stableMemoryLoadInt32") as p), [e] ->
+  | OtherPrim ("stableMemoryLoadNat32" | "stableMemoryLoadInt32" as p), [e] ->
     (SR.UnboxedWord32 Type.(if p = "stableMemoryLoadNat32" then Nat32 else Int32)),
     compile_exp_as env ae (SR.UnboxedWord64 Type.Nat64) e ^^
     StableMemoryInterface.load_word32 env
 
-  | OtherPrim (("stableMemoryStoreNat32" | "stableMemoryStoreInt32") as p), [e1; e2] ->
+  | OtherPrim ("stableMemoryStoreNat32" | "stableMemoryStoreInt32" as p), [e1; e2] ->
     SR.unit,
     compile_exp_as env ae (SR.UnboxedWord64 Type.Nat64) e1 ^^
     compile_exp_as env ae
@@ -12241,7 +12240,7 @@ and compile_prim_invocation (env : E.t) ae p es at =
       e2 ^^
     StableMemoryInterface.store_word32 env
 
-  | OtherPrim (("stableMemoryLoadNat8" | "stableMemoryLoadInt8") as p), [e] ->
+  | OtherPrim ("stableMemoryLoadNat8" | "stableMemoryLoadInt8" as p), [e] ->
     let ty = Type.(if p = "stableMemoryLoadNat8" then Nat8 else Int8) in
     SR.UnboxedWord32 ty,
     compile_exp_as env ae (SR.UnboxedWord64 Type.Nat64) e ^^
@@ -12250,7 +12249,7 @@ and compile_prim_invocation (env : E.t) ae p es at =
 
   (* Other prims, binary *)
 
-  | OtherPrim (("stableMemoryStoreNat8" | "stableMemoryStoreInt8") as p), [e1; e2] ->
+  | OtherPrim ("stableMemoryStoreNat8" | "stableMemoryStoreInt8" as p), [e1; e2] ->
     let ty = Type.(if p = "stableMemoryStoreNat8" then Nat8 else Int8) in
     SR.unit,
     compile_exp_as env ae (SR.UnboxedWord64 Type.Nat64) e1 ^^
@@ -12258,14 +12257,14 @@ and compile_prim_invocation (env : E.t) ae p es at =
     TaggedSmallWord.lsb_adjust ty ^^
     StableMemoryInterface.store_word8 env
 
-  | OtherPrim (("stableMemoryLoadNat16" | "stableMemoryLoadInt16") as p), [e] ->
+  | OtherPrim ("stableMemoryLoadNat16" | "stableMemoryLoadInt16" as p), [e] ->
     let ty = Type.(if p = "stableMemoryLoadNat16" then Nat16 else Int16) in
     SR.UnboxedWord32 ty,
     compile_exp_as env ae (SR.UnboxedWord64 Type.Nat64) e ^^
     StableMemoryInterface.load_word16 env ^^
     TaggedSmallWord.msb_adjust ty
 
-  | OtherPrim (("stableMemoryStoreNat16" | "stableMemoryStoreInt16") as p), [e1; e2] ->
+  | OtherPrim ("stableMemoryStoreNat16" | "stableMemoryStoreInt16" as p), [e1; e2] ->
     let ty = Type.(if p = "stableMemoryStoreNat16" then Nat16 else Int16) in
     SR.unit,
     compile_exp_as env ae (SR.UnboxedWord64 Type.Nat64) e1 ^^
@@ -12273,12 +12272,12 @@ and compile_prim_invocation (env : E.t) ae p es at =
     TaggedSmallWord.lsb_adjust ty ^^
     StableMemoryInterface.store_word16 env
 
-  | OtherPrim (("stableMemoryLoadNat64" | "stableMemoryLoadInt64") as p), [e] ->
+  | OtherPrim ("stableMemoryLoadNat64" | "stableMemoryLoadInt64" as p), [e] ->
     (SR.UnboxedWord64 Type.(if p = "stableMemoryLoadNat64" then Nat64 else Int64)),
     compile_exp_as env ae (SR.UnboxedWord64 Type.Nat64) e ^^
     StableMemoryInterface.load_word64 env
 
-  | OtherPrim (("stableMemoryStoreNat64" | "stableMemoryStoreInt64") as p), [e1; e2] ->
+  | OtherPrim ("stableMemoryStoreNat64" | "stableMemoryStoreInt64" as p), [e1; e2] ->
     SR.unit,
     compile_exp_as env ae (SR.UnboxedWord64 Type.Nat64) e1 ^^
     compile_exp_as env ae
@@ -12403,7 +12402,7 @@ and compile_prim_invocation (env : E.t) ae p es at =
     const_sr SR.Vanilla (Text.to_blob env)
 
   (* textual to bytes *)
-  | BlobOfIcUrl, [_] ->
+  | (OtherPrim "decode_principal" | BlobOfIcUrl), [_] ->
     const_sr SR.Vanilla (E.call_import env "rts" "blob_of_principal")
   (* The other direction *)
   | IcUrlOfBlob, [_] ->
@@ -12750,7 +12749,7 @@ and compile_exp_with_hint (env : E.t) ae sr_hint exp =
       add_cycles
   | ActorE (ds, fs, _, _) ->
     fatal "Local actors not supported by backend"
-  | NewObjE (Type.(Object | Module | Memory) as _sort, fs, _) ->
+  | NewObjE (Type.(Object | Module | Memory), fs, _) ->
     (*
     We can enable this warning once we treat everything as static that
     mo_frontend/static.ml accepts, including _all_ literals.