+++ src/lowering/desugar.ml
@@ -482,7 +482,7 @@ and call_system_func_opt name es obj_typ =
            let timer =
              blockE
                [ expD T.(callE (varE (var id.it note)) [Any]
-                   (varE (var "@set_global_timer" T.global_timer_set_type))) ]
+                   (varE (var "@set_global_timer" global_timer_set_type))) ]
                (unitE()) in
            { timer with at }
         | "heartbeat" ->
@@ -1272,8 +1272,10 @@ and transform_import (i : S.import) : Ir.dec list =
       varE (var (id_of_full_path fp) t)
     | S.PrimPath ->
       varE (var (id_of_full_path "@prim") t)
-    | S.IDLPath (fp, canister_id) ->
+    | S.IDLPath (fp, Either.Right canister_id) ->
       primE (I.ActorOfIdBlob t) [blobE canister_id]
+    | S.IDLPath (fp, Either.Left envvar) ->
+      primE (I.ActorOfIdBlob t) T.[callE (varE (var "@envvar_principal" (Func (Local, Returns, [], [Prim Text], [Prim Blob])))) [] (textE envvar)]
     | S.ImportedValuePath path ->
        if !Mo_config.Flags.blob_import_placeholders then
          raise (Invalid_argument ("blob import placeholder"))