+++ src/mo_def/syntax.ml
@@ -16,7 +16,7 @@ type lib_path = {package : string option; path : string}
 type resolved_import =
   | Unresolved
   | LibPath of lib_path
-  | IDLPath of (string * string) (* filepath * bytes *)
+  | IDLPath of (string * (string, string) Either.t) (* filepath * envvar/bytes *)
   | ImportedValuePath of string
   | PrimPath (* the built-in prim module *)