+++ src/pipeline/resolve_import.mli
@@ -8,7 +8,7 @@ module RIM : Map.S with type key = Syntax.resolved_import
 
 type actor_idl_path = string option
 type package_urls = string Flags.M.t
-type actor_aliases = string Flags.M.t
+type actor_aliases = (string * string, string * string option) Either.t Flags.M.t
 
 type resolved_imports = Syntax.resolved_import Source.phrase list
 
@@ -23,7 +23,8 @@ type flags = {
 
 type package_map = filepath Flags.M.t
 type blob = string
-type aliases = blob Flags.M.t
+type envvar = string
+type aliases = (envvar * filepath, blob * filepath option) Either.t Flags.M.t
 type resolved_flags = {
   packages : package_map;
   aliases : aliases;